Adding an observable member into class UartDriverRaritan

+++ src/spi/raritan/RaritanUartInputObservable.cpp
@@ -11,7 +11,7 @@ RaritanUartInputObservable::~RaritanUartInputObservable() {
 
 void RaritanUartInputObservable::genRandomData() {
 	unsigned char asyncData[20];
-	size_t dataLen = static_cast<size_t>(1+rand()%sizeof(asyncData));
+	size_t dataLen = static_cast<size_t>(1+static_cast<unsigned int>(rand())%sizeof(asyncData));
 	std::cout << "Observable is generating random buffer of " << dataLen << " bytes" << std::endl;
 	for (unsigned int bytePos=0; bytePos<dataLen; bytePos++) {
 		asyncData[bytePos]=static_cast<unsigned char>(rand() % 256);
